gameplay: Gp_UpdateAttachCombo without matching hacks
The pins and barriers stood in for a switch over the item id with a block of
locals per case, direct Gp_StateC08 access, and a table pointer taken before the
level is computed. GpStateC08.field_D is signed, as its other readers already
cast it.

+++ src/gameplay/3A34.c
@@ -6456,7 +6456,7 @@ u32 Gp_ComputeDamage(u32 arg0, u32 arg1, s32 arg2, s32 arg3)
         dmg = tmp * mult >> 16;
 
         if (flag != 0) {
-            extra = (s8)Gp_StateC08.field_D;
+            extra = Gp_StateC08.field_D;
             if (extra != 0) {
                 dmg = dmg * D_80113D0C[(extra / 16 - 1) * 2 + (s8)(extra % 16)][0] / 100;
             }
@@ -6616,7 +6616,7 @@ s32 Gp_RollEnemyChance(GpEnemy* arg0, u32 arg1, s32 arg2)
         chance <<= 1;
     }
 
-    extra = (s8)Gp_StateC08.field_D;
+    extra = Gp_StateC08.field_D;
     if (extra != 0) {
         chance = chance * D_80113D0C[(extra / 16 - 1) * 2 + (s8)(extra % 16)][1] / 100;
     }
